backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
PyPDF2==3.0.1
psycopg2-binary>=2.9.11
boto3==1.34.84
chromadb>=0.5.0
sentence-transformers==2.6.1
python-dotenv==1.0.1
